refactor(dsl): split root and schema entrypoints
