+++ dsl/src/main/kotlin/me/liam/microsmith/dsl/core/MicrosmithExtension.kt
@@ -1,0 +1,6 @@
+package me.liam.microsmith.dsl.core
+
+/**
+ * Marker interface for extensions declared at the root `microsmith { ... }` scope.
+ */
+interface MicrosmithExtension : ModelExtension
