+++ dsl/src/main/kotlin/me/liam/microsmith/dsl/core/ModelExtension.kt
@@ -1,0 +1,6 @@
+package me.liam.microsmith.dsl.core
+
+/**
+ * Marker interface for typed extension payloads stored on a [MicrosmithModel].
+ */
+interface ModelExtension
